requirements.txt
asgiref==3.3.4
Django==3.2.4
django-ex00==0.1
gunicorn==20.1.0
Pillow==8.2.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
uWSGI==2.0.19.1
